api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
boto3==1.34.34
botocore==1.34.34
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1
python-multipart==0.0.6
cryptography==42.0.0
aiofiles==23.2.1
httpx==0.26.0
tenacity==8.2.3
prometheus-client==0.19.0
structlog==24.1.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
